requirements.txt
pyrogram
TgCrypto
py-tgcalls==0.5.5
python-dotenv
youtube_search_python==1.4.6
requests
aiohttp[speedups]
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg
Pillow
future
psutil
wget
yt_dlp
bs4
dnspython
SQLAlchemy
motor
httpx==0.13.3
heroku3
gitpython
